Extract Pkcs8KeyPairParser; stop OpenSSHKeyFile extending PKCS8KeyFile
PKCS8KeyFile had no state - it existed only to override readKeyPair() with
PEM/PKCS8/PKCS1 parsing, and OpenSSHKeyFile extended it purely to reuse
that parsing while adding companion-public-key support. Inheritance for
implementation reuse.

Move the parsing into a KeyPairParser strategy (Pkcs8KeyPairParser).
PKCS8KeyFile and OpenSSHKeyFile now both extend BaseFileKeyProvider
directly and delegate readKeyPair() to the parser. The parser is a pure
function of (Resource, PasswordFinder), independently testable, with no
provider state threaded through instance fields.

No behaviour change: the parsing code moved verbatim.
